+++ src/test/java/uk/gov/legislation/api/test/LanguageRecognitionTest.java
@@ -6,7 +6,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.web.servlet.LocaleResolver;
 import org.springframework.web.servlet.i18n.AcceptHeaderLocaleResolver;
-import uk.gov.legislation.endpoints.Application;
+import uk.gov.legislation.Application;
 
 import java.util.Locale;
