+++ src/test/java/uk/gov/legislation/api/test/MetadataExtractionTest.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import uk.gov.legislation.api.responses.DocumentMetadata;
 import uk.gov.legislation.converters.DocumentMetadataConverter;
-import uk.gov.legislation.endpoints.Application;
+import uk.gov.legislation.Application;
 import uk.gov.legislation.transform.simple.Contents;
 import uk.gov.legislation.transform.simple.Simplify;
