+++ src/test/java/uk/gov/legislation/api/test/Tests.java
@@ -2,7 +2,7 @@ package uk.gov.legislation.api.test;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import uk.gov.legislation.endpoints.Application;
+import uk.gov.legislation.Application;
 
 @SpringBootTest(classes = Application.class)
 class Tests {
