+++ src/test/java/uk/gov/legislation/api/test/TocTest.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import uk.gov.legislation.api.responses.TableOfContents;
 import uk.gov.legislation.converters.TableOfContentsConverter;
-import uk.gov.legislation.endpoints.Application;
+import uk.gov.legislation.Application;
 import uk.gov.legislation.transform.simple.Contents;
 import uk.gov.legislation.transform.simple.Simplify;
 import uk.gov.legislation.util.UpToDate;
