+++ src/test/java/uk/gov/legislation/api/test/UnappliedEffectsTest.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import uk.gov.legislation.api.responses.Effect;
 import uk.gov.legislation.converters.EffectsFeedConverter;
-import uk.gov.legislation.endpoints.Application;
+import uk.gov.legislation.Application;
 import uk.gov.legislation.transform.simple.Metadata;
 import uk.gov.legislation.transform.simple.Simplify;
 import uk.gov.legislation.util.Effects;
